Reorganized repository into a package
* added a notebook initialization script (`nbinit`)
* added .gitignore
* updated xarray timedelta decoding behavior
* updated basic notebook template
