+++ .gitignore
@@ -1,0 +1,3 @@
+.idea
+*.egg-info
+__pycache__
