+++ tests/test_data/input_timeslice_test.yml
@@ -1,0 +1,77 @@
+# Runtime yaml configuration file template for the MDTF-diagnostics package
+# Create a copy of this file for personal use, and pass it to the framework
+# with the -f | --configfile flag
+
+# List of POD(s) to run
+pod_list:
+#  - "example_multicase"
+  - "Wheeler_Kiladis"
+#  - "MJO_suite"
+#  - "precip_diurnal_cycle"
+
+
+# Case list entries (must be unique IDs for each simulation)
+case_list:
+  "timeslice_mdtfv3" :
+    model: "CESM"
+    convention: "CESM"
+    startdate: "19950101000000"
+    enddate:   "19961231000000"
+
+### Data location settings ###
+# Required: full or relative path to ESM-intake catalog header file
+DATA_CATALOG: "/glade/u/home/bundy/diag/mdtf/catalogs/esm_catalog_CESM_timeslice_mdtfv3.20241107.json"
+
+# Optional: Parent directory containing observational data used by individual PODs.
+# If defined, the framework assumes observational data is in OBS_DATA_ROOT/[POD_NAME]
+OBS_DATA_ROOT: "/glade/work/bundy/mdtf/inputdata/obs_data"
+
+# Required: Working directory location. Temporary files are written here.
+# Final output is also written here if the OUTPUT_DIR is not defined.
+WORK_DIR: "/glade/work/bundy/mdtf/outdir/20250403"
+
+# Optional: Location to write final output if you don't want it in the wkdir
+OUTPUT_DIR: ""
+
+### Environment Settings ###
+# Required: Location of the Anaconda/miniconda installation to use for managing
+# dependencies (path returned by running `conda info --base`.)
+#conda_root: "/glade/u/home/bundy/miniconda2"
+conda_root: "/glade/work/bundy/miniconda2/"
+
+
+# Optional: Directory containing the framework-specific conda environments. This should
+# be equal to the "--env_dir" flag passed to conda_env_setup.sh. If left
+# blank, the framework will look for its environments in conda_root/envs
+conda_env_root: "/glade/work/bundy/miniconda2/envs.MDTF.20241107/"
+
+# Location of micromamba executable; REQUIRED if using micromamba
+micromamba_exe: ""
+
+### Data type settings ###
+# set to true to handle data files > 4 GB
+large_file: False
+
+### Output Settings ###
+# Set to true to have PODs save postscript figures in addition to bitmaps.
+save_ps: False
+# If true, leave pp data in OUTPUT_DIR after preprocessing; if false, delete pp data after PODs
+# run to completion
+save_pp_data: True
+# Set to true to perform data translation; default is True:
+translate_data: True
+# Set to true to save HTML and bitmap plots in a .tar file.
+make_variab_tar: False
+# Set to true to overwrite results in OUTPUT_DIR; otherwise results saved
+# under a unique name.
+overwrite: True
+# Generate html output for multiple figures per case
+"make_multicase_figure_html": False
+### Developer settings ###
+# Set to True to run the preprocessor
+run_pp: True
+# Additional custom preprocessing scripts to run on data
+# place these scripts in the MDTF-diagnostics/user_scripts directory
+# The framework will run the specified scripts whether run_pp is set to True or False
+user_pp_scripts:
+  - ""
